+++ chart/src/commonMain/kotlin/render/BarChartRenderer.kt
@@ -1,6 +1,8 @@
 package com.juul.krayon.chart.render
 
 import com.juul.krayon.canvas.Canvas
+import com.juul.krayon.canvas.Color
+import com.juul.krayon.canvas.Paint
 import com.juul.krayon.chart.data.ClusteredDataSet
 import com.juul.krayon.chart.render.components.AxisRenderer
 import com.juul.krayon.chart.render.components.BarRenderer
@@ -27,7 +29,7 @@ public open class BarChartRenderer(
         val dataAreaTop = AXIS_PAD
         val dataAreaRight = canvas.width - AXIS_PAD
         val dataAreaBottom = canvas.height - AXIS_PAD
-
+        canvas.drawRect(0f, 0f, canvas.width, canvas.height, canvas.buildPaint(Paint.Fill(Color.white)))
         barRenderer.render(BarRenderer.Specification(dataAreaLeft, dataAreaTop, dataAreaRight, dataAreaBottom, data), canvas)
         xAxisRenderer.render(AxisRenderer.Specification(dataAreaLeft, dataAreaBottom, dataAreaRight), canvas)
         yAxisRenderer.render(AxisRenderer.Specification(dataAreaLeft, dataAreaBottom, dataAreaTop), canvas)
