+++ sample/src/jsMain/resources/index.html
@@ -5,7 +5,7 @@
     <title>Hello, Kotlin/JS!</title>
 </head>
 <body>
-    <canvas id="canvas"></canvas>
+    <canvas id="canvas" width="854" height="480"></canvas>
 </body>
 <script src="sample.js"></script>
 </html>
